backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Agent 对话系统 — 后端"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "redis[asyncio]>=5.2.0",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.8.0",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "openai>=1.60.0",
    "pymupdf>=1.24.0",
    "python-multipart>=0.0.9",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "langgraph>=0.2.28",
    "langchain-openai>=0.2.0",
    "pymilvus>=2.4.0",
    "alibabacloud-docmind-api20220711>=1.3.0",
    "alibabacloud-tea-openapi>=0.3.0",
    "alibabacloud-tea-util>=0.3.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = false
